upload_file_api/upload_file/views.py
from rest_framework import generics, status
from rest_framework.response import Response
from .tasks import process_file_task

from .models import File
from .serializers import FileSerializer


class FileListAPIView(generics.ListAPIView):
    queryset = File.objects.all()
    serializer_class = FileSerializer


class FileCreateAPIView(generics.CreateAPIView):
    queryset = File.objects.all()
    serializer_class = FileSerializer

    def create(self, request, *args, **kwargs):
        serializer = self.serializer_class(data=request.data)
        if serializer.is_valid():
            file_instance = serializer.save()
            process_file_task.delay(file_instance.id)
            return Response(serializer.data, status=status.HTTP_201_CREATED)
        else:
            return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
